+++ render/cw1/point_cloud.hpp
@@ -11,7 +11,7 @@
 struct PointCloud {
     std::vector<glm::vec3> positions;
     std::vector<glm::vec3> colors;
-    std::vector<float> point_size;
+    std::vector<int> point_size;
 
     void set_color(glm::vec3 const& color);
     void set_size(unsigned int const& size);
@@ -32,7 +32,7 @@ struct LineBuffer {
 };
 
 
-PointBuffer create_pointcloud_buffers(std::vector<glm::vec3> positions, std::vector<glm::vec3> color, std::vector<float> scale,
+PointBuffer create_pointcloud_buffers(std::vector<glm::vec3> positions, std::vector<glm::vec3> color, std::vector<int> scale,
                              labutils::VulkanContext const& window, labutils::Allocator const& allocator);
 
 LineBuffer create_index_buffer(std::vector<uint32_t> const&, std::vector<glm::vec3> const&, labutils::VulkanContext const& window, labutils::Allocator const& allocator);
